Changed GroundDetection to GroundDetection2D.
There is a weird bug where after setting velocity, in the next frame, Physics
still sees the object as a contact despite the objects being visually seperated.
This causes GroundDetection2D to set IsOnGround to true despite the fact that the
player object has been moving along a new upward velocity for a frame. This
causes the Y component of the move velocity to be zero'd out the frame after it
is set to JumpVelocity, cancelling the jump just after it gets off the ground.
I need to talk to @jodavis about this.